+++ src/widgets/splits/SplitInput.cpp
@@ -309,8 +309,6 @@ void SplitInput::openEmotePopup()
             });
     }
 
-    this->emotePopup_->resize(int(300 * this->emotePopup_->scale()),
-                              int(500 * this->emotePopup_->scale()));
     this->emotePopup_->loadChannel(this->split_->getChannel());
     this->emotePopup_->show();
     this->emotePopup_->raise();